Moved basicSchema and messageSchema into src/models folder
